+++ NEWS.md
@@ -26,7 +26,7 @@ colors are very distinct from each other.
 * `overlay_function` convenience function. Example usage: add a Gaussian (or any
 distribution) density curve to a plot made with `mcmc_hist`.
 * `mcmc_recover_scatter`, which is similar to `mcmc_recover_intervals` but plots
-estimated vs "true" values as a scatterplot.
+estimated vs "true" values as a scatterplot. (#81)
 * New PPC category __Discrete__ with functions: 
     - `ppc_rootogram` for use with models for count data. Thanks to @paul-buerkner.
 (#28)